% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppenderRedis.R
\name{filter_custom_fields}
\alias{filter_custom_fields}
\title{Filter Custom Fields}
\usage{
filter_custom_fields(event)
}
\arguments{
\item{event}{(\link[lgr:LogEvent]{lgr::LogEvent})\cr
Log event.}
}
\description{
Filters custom fields from log events.
}
