% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\name{rsh}
\alias{rsh}
\title{Synctatic Sugar for Rush Controller Construction}
\usage{
rsh(instance_id = NULL, config = NULL, ....)
}
\arguments{
\item{instance_id}{(\code{character(1)})\cr
Identifier of the rush network.
Controller and workers must have the same instance id.
Keys in Redis are prefixed with the instance id.}

\item{config}{(\link[redux:redis_config]{redux::redis_config})\cr
Redis configuration options.
See \link[redux:redis_config]{redux::redis_config} for details.}
}
\description{
Function to construct a \link{Rush} controller.
}
\examples{
rsh(instance_id = "benchmark")
}
