% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker_loops.R
\name{run_worker}
\alias{run_worker}
\title{Worker Function}
\usage{
run_worker(
  worker_loop,
  instance_id,
  config,
  host,
  worker_id,
  heartbeat_period,
  heartbeat_expire,
  args
)
}
\arguments{
\item{instance_id}{(\code{character(1)})\cr
Identifier of the rush instance.}

\item{config}{(\link[redux:redis_config]{redux::redis_config})\cr
Redis configuration.}

\item{host}{(\code{character(1)})\cr
Local or remote host.}

\item{worker_id}{(\code{character(1)})\cr
Identifier of the worker.}

\item{heartbeat_period}{(\code{numeric(1)})\cr
Period of the heartbeat.}

\item{heartbeat_expire}{(\code{numeric(1)})\cr
Expiration of the heartbeat.}

\item{fun}{(\code{function})\cr
Function to be executed.}
}
\description{
Called on the worker.
Initializes the rush worker instance and invokes the worker loop.
}
