% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rush-package.R
\docType{package}
\name{rush-package}
\alias{rush}
\alias{rush-package}
\title{rush: R One-Liners from the Shell}
\description{
Run R expressions and create plots directly from the shell. Each invocation assembles a self-contained script that is executed with the 'ir' command-line tool, which resolves the script's package dependencies on the fly.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jeroenjanssens.github.io/rush/}
  \item \url{https://github.com/jeroenjanssens/rush}
  \item Report bugs at \url{https://github.com/jeroenjanssens/rush/issues}
}

}
\author{
\strong{Maintainer}: Jeroen Janssens \email{jeroen@jeroenjanssens.com} (\href{https://orcid.org/0000-0002-5878-8924}{ORCID})

}
\keyword{internal}
