% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rush_plan.R
\name{rush_available}
\alias{rush_available}
\title{Rush Available}
\usage{
rush_available()
}
\value{
\code{logical(1)}
}
\description{
Returns \code{TRUE} if a redis config file (\code{\link[redux:redis_config]{redux::redis_config()}}) has been set by \code{\link[=rush_plan]{rush_plan()}}.
}
