% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rush_plan.R
\name{rush_config}
\alias{rush_config}
\title{Get Rush Config}
\usage{
rush_config()
}
\value{
\code{\link[redux:redis_config]{redux::redis_config()}}\cr
A redis config file.
}
\description{
Returns the redis config file (\link[redux:redis_config]{redux::redis_config}) that was set by \code{\link[=rush_plan]{rush_plan()}}.
}
