% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_large_object.R
\name{store_large_object}
\alias{store_large_object}
\title{Store Large Objects}
\usage{
store_large_object(obj, path)
}
\arguments{
\item{obj}{(\code{any})\cr
Object to store.}

\item{path}{(\code{character(1)})\cr
Path to store the object.}
}
\value{
\code{list()} of class \code{"rush_large_object"} with the name and path of the stored object.
}
\description{
Store large objects to disk and return a reference to the object.
}
\examples{
obj = list(a = 1, b = 2)
rush_large_object = store_large_object(obj, tempdir())
}
