% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker_loops.R
\name{worker_loop_default}
\alias{worker_loop_default}
\title{Single Task Worker Loop}
\usage{
worker_loop_default(fun, rush)
}
\arguments{
\item{fun}{(\code{function})\cr
Function to be executed.}

\item{rush}{(\link{RushWorker})\cr
Rush worker instance.}
}
\description{
Worker loop that pops a single task from the queue, executes the function and pushes the results.
}
