% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heartbeat_loops.R
\name{fun_heartbeat}
\alias{fun_heartbeat}
\title{Heartbeat Loop}
\usage{
fun_heartbeat(instance_id, config, worker_id, period, expire, pid)
}
\arguments{
\item{instance_id}{(\code{character(1)})\cr
Identifier of the rush instance.}

\item{config}{(\link[redux:redis_config]{redux::redis_config})\cr
Redis configuration.}

\item{worker_id}{(\code{character(1)})\cr
Identifier of the worker.}

\item{period}{(\code{numeric(1)})\cr
Period of the heartbeat.}

\item{expire}{(\code{numeric(1)})\cr
Expiration of the heartbeat.}

\item{pid}{(\code{numeric(1)})\cr
Process ID of the worker.}
}
\description{
Heartbeat loop that updates the heartbeat key and checks for kill key.
}
