% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker_loops.R
\name{fun_loop}
\alias{fun_loop}
\title{Single Task Worker Loop}
\usage{
fun_loop(fun, rush)
}
\arguments{
\item{fun}{(\code{function})\cr
Function to be executed.}

\item{rush}{(\code{character(1)})\cr
Rush worker instance.}
}
\description{
Worker loop that pops a single task from the queue, executes the function and pushes the results.
}
