% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_worker.R
\name{run_worker}
\alias{run_worker}
\title{Run a Worker.}
\usage{
run_worker(
  worker_loop,
  instance_id,
  config,
  host,
  worker_id,
  heartbeat_period,
  heartbeat_expire,
  lgr_thresholds,
  args
)
}
\arguments{
\item{worker_loop}{(\code{function})\cr
Worker loop to be executed e.g. \link{fun_loop}.}

\item{instance_id}{(\code{character(1)})\cr
Identifier of the rush instance.}

\item{config}{(\link[redux:redis_config]{redux::redis_config})\cr
Redis configuration.}

\item{host}{(\code{character(1)})\cr
Local or remote host.}

\item{worker_id}{(\code{character(1)})\cr
Identifier of the worker.}

\item{heartbeat_period}{(\code{numeric(1)})\cr
Period of the heartbeat.}

\item{heartbeat_expire}{(\code{numeric(1)})\cr
Expiration of the heartbeat.}

\item{lgr_thresholds}{(named \code{character()} or \code{numeric()})\cr
Logger thresholds.}
}
\description{
Runs a worker.
The function initializes the \link{RushWorker} instance and invokes the worker loop.
}
