% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{rush-package}
\alias{rush}
\alias{rush-package}
\title{rush: Parallele Programmierung}
\description{
Parallel programmierung with future and redis.
}
\author{
\strong{Maintainer}: Marc Becker \email{marcbecker@posteo.de} (\href{https://orcid.org/0000-0002-8115-0400}{ORCID})

}
