% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{rush-package}
\alias{rush}
\alias{rush-package}
\title{rush: Rapid Parallel and Distributed Computing}
\description{
Parallel computing with a network of local and remote workers. Fast exchange of results between the workers through a 'Redis' database. Key features include task queues, local caching, and sophisticated error handling.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mlr-org/rush}
  \item Report bugs at \url{https://github.com/mlr-org/rush/issues}
}

}
\author{
\strong{Maintainer}: Marc Becker \email{marcbecker@posteo.de} (\href{https://orcid.org/0000-0002-8115-0400}{ORCID}) [copyright holder]

}
