% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{rush-package}
\alias{rush}
\alias{rush-package}
\title{rush: Rapid Asynchronous and Distributed Computing}
\description{
Package to tackle large-scale problems asynchronously across a distributed network. Employing a database centric model, rush enables workers to communicate tasks and their results over a shared 'Redis' database. Key features include low task overhead, efficient caching, and robust error handling. The package powers the asynchronous optimization algorithms in the 'bbotk' and 'mlr3tuning' packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rush.mlr-org.com}
  \item \url{https://github.com/mlr-org/rush}
  \item Report bugs at \url{https://github.com/mlr-org/rush/issues}
}

}
\author{
\strong{Maintainer}: Marc Becker \email{marcbecker@posteo.de} (\href{https://orcid.org/0000-0002-8115-0400}{ORCID}) [copyright holder]

}
