% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rush_plan.R
\name{rush_config}
\alias{rush_config}
\title{Get Rush Config}
\usage{
rush_config()
}
\value{
\code{list()} with the stored configuration.
}
\description{
Returns the rush config that was set by \code{\link[=rush_plan]{rush_plan()}}.
}
\examples{
# This example is not executed since Redis must be installed
\donttest{
  config_local = redux::redis_config()
  rush_plan(config = config_local, n_workers = 2)
  rush_config()
}
}
