% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rush_plan.R
\name{rush_plan}
\alias{rush_plan}
\title{Create Rush Plan}
\usage{
rush_plan(config, n_workers)
}
\arguments{
\item{config}{\code{\link[redux:redis_config]{redux::redis_config()}}\cr
A redis config file.}
}
\description{
Stores a redis config file (\code{\link[redux:redis_config]{redux::redis_config()}}) that can be used by \code{Rush} to connect to a redis server.
}
