% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RushWorker.R
\name{RushWorker}
\alias{RushWorker}
\title{Rush Worker}
\value{
Object of class \link[R6:R6Class]{R6::R6Class} and \code{RushWorker}.
}
\description{
\link{RushWorker} inherits all methods from \link{Rush}.
Upon initialization, the worker registers itself in the Redis database as a running worker.
This class is usually not constructed directly by the user.

In addition to the inherited methods, the worker provides methods that require a worker identity:
\itemize{
\item \verb{$pop_task()}: Pop a task from the queue and mark it as running.
\item \verb{$push_running_tasks(xss)}: Create running tasks evaluated by the worker.
\item \verb{$finish_tasks(keys, yss)}: Save the output of tasks and mark them as finished.
\item \verb{$fail_tasks(keys, conditions)}: Mark tasks as failed and optionally save the condition objects.
}
}
\section{Super class}{
\code{\link[rush:Rush]{Rush}} -> \code{RushWorker}
}
\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{worker_id}}{(\code{character(1)})\cr
Identifier of the worker.}

    \item{\code{profile}}{(\code{character(1)})\cr
Name of the \code{mirai} compute profile the worker runs on.
\code{NULL} if the worker runs on the default compute profile.}

    \item{\code{heartbeat}}{(\code{callr::r_bg})\cr
Background process for the heartbeat.}
  }
  \if{html}{\out{</div>}}
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{terminated}}{(\code{logical(1)})\cr
Whether to shutdown the worker.
Used in the worker loop to determine whether to continue.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-RushWorker-initialize}{\code{RushWorker$new()}}
    \item \href{#method-RushWorker-pop_task}{\code{RushWorker$pop_task()}}
    \item \href{#method-RushWorker-push_running_tasks}{\code{RushWorker$push_running_tasks()}}
    \item \href{#method-RushWorker-finish_tasks}{\code{RushWorker$finish_tasks()}}
    \item \href{#method-RushWorker-fail_tasks}{\code{RushWorker$fail_tasks()}}
    \item \href{#method-RushWorker-set_terminated}{\code{RushWorker$set_terminated()}}
    \item \href{#method-RushWorker-clone}{\code{RushWorker$clone()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="rush" data-topic="Rush" data-id="detect_lost_workers"><a href='../../rush/html/Rush.html#method-Rush-detect_lost_workers'><code>Rush$detect_lost_workers()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="rush" data-topic="Rush" data-id="empty_queue"><a href='../../rush/html/Rush.html#method-Rush-empty_queue'><code>Rush$empty_queue()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="rush" data-topic="Rush" data-id="fetch_failed_tasks"><a href='../../rush/html/Rush.html#method-Rush-fetch_failed_tasks'><code>Rush$fetch_failed_tasks()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="rush" data-topic="Rush" data-id="fetch_finished_tasks"><a href='../../rush/html/Rush.html#method-Rush-fetch_finished_tasks'><code>Rush$fetch_finished_tasks()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="rush" data-topic="Rush" data-id="fetch_new_tasks"><a href='../../rush/html/Rush.html#method-Rush-fetch_new_tasks'><code>Rush$fetch_new_tasks()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="rush" data-topic="Rush" data-id="fetch_queued_tasks"><a href='../../rush/html/Rush.html#method-Rush-fetch_queued_tasks'><code>Rush$fetch_queued_tasks()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="rush" data-topic="Rush" data-id="fetch_running_tasks"><a href='../../rush/html/Rush.html#method-Rush-fetch_running_tasks'><code>Rush$fetch_running_tasks()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="rush" data-topic="Rush" data-id="fetch_tasks"><a href='../../rush/html/Rush.html#method-Rush-fetch_tasks'><code>Rush$fetch_tasks()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="rush" data-topic="Rush" data-id="fetch_tasks_with_state"><a href='../../rush/html/Rush.html#method-Rush-fetch_tasks_with_state'><code>Rush$fetch_tasks_with_state()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="rush" data-topic="Rush" data-id="format"><a href='../../rush/html/Rush.html#method-Rush-format'><code>Rush$format()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="rush" data-topic="Rush" data-id="is_failed_task"><a href='../../rush/html/Rush.html#method-Rush-is_failed_task'><code>Rush$is_failed_task()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="rush" data-topic="Rush" data-id="is_running_task"><a href='../../rush/html/Rush.html#method-Rush-is_running_task'><code>Rush$is_running_task()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="rush" data-topic="Rush" data-id="print"><a href='../../rush/html/Rush.html#method-Rush-print'><code>Rush$print()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="rush" data-topic="Rush" data-id="print_log"><a href='../../rush/html/Rush.html#method-Rush-print_log'><code>Rush$print_log()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="rush" data-topic="Rush" data-id="push_failed_tasks"><a href='../../rush/html/Rush.html#method-Rush-push_failed_tasks'><code>Rush$push_failed_tasks()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="rush" data-topic="Rush" data-id="push_finished_tasks"><a href='../../rush/html/Rush.html#method-Rush-push_finished_tasks'><code>Rush$push_finished_tasks()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="rush" data-topic="Rush" data-id="push_tasks"><a href='../../rush/html/Rush.html#method-Rush-push_tasks'><code>Rush$push_tasks()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="rush" data-topic="Rush" data-id="read_hash"><a href='../../rush/html/Rush.html#method-Rush-read_hash'><code>Rush$read_hash()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="rush" data-topic="Rush" data-id="read_hashes"><a href='../../rush/html/Rush.html#method-Rush-read_hashes'><code>Rush$read_hashes()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="rush" data-topic="Rush" data-id="read_log"><a href='../../rush/html/Rush.html#method-Rush-read_log'><code>Rush$read_log()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="rush" data-topic="Rush" data-id="reconnect"><a href='../../rush/html/Rush.html#method-Rush-reconnect'><code>Rush$reconnect()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="rush" data-topic="Rush" data-id="reset"><a href='../../rush/html/Rush.html#method-Rush-reset'><code>Rush$reset()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="rush" data-topic="Rush" data-id="reset_cache"><a href='../../rush/html/Rush.html#method-Rush-reset_cache'><code>Rush$reset_cache()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="rush" data-topic="Rush" data-id="start_local_workers"><a href='../../rush/html/Rush.html#method-Rush-start_local_workers'><code>Rush$start_local_workers()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="rush" data-topic="Rush" data-id="start_remote_workers"><a href='../../rush/html/Rush.html#method-Rush-start_remote_workers'><code>Rush$start_remote_workers()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="rush" data-topic="Rush" data-id="start_workers"><a href='../../rush/html/Rush.html#method-Rush-start_workers'><code>Rush$start_workers()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="rush" data-topic="Rush" data-id="stop_workers"><a href='../../rush/html/Rush.html#method-Rush-stop_workers'><code>Rush$stop_workers()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="rush" data-topic="Rush" data-id="tasks_with_state"><a href='../../rush/html/Rush.html#method-Rush-tasks_with_state'><code>Rush$tasks_with_state()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="rush" data-topic="Rush" data-id="wait_for_tasks"><a href='../../rush/html/Rush.html#method-Rush-wait_for_tasks'><code>Rush$wait_for_tasks()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="rush" data-topic="Rush" data-id="wait_for_workers"><a href='../../rush/html/Rush.html#method-Rush-wait_for_workers'><code>Rush$wait_for_workers()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="rush" data-topic="Rush" data-id="worker_script"><a href='../../rush/html/Rush.html#method-Rush-worker_script'><code>Rush$worker_script()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="rush" data-topic="Rush" data-id="write_hashes"><a href='../../rush/html/Rush.html#method-Rush-write_hashes'><code>Rush$write_hashes()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RushWorker-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-RushWorker-initialize}{}}}
\subsection{\code{RushWorker$new()}}{
  Creates a new instance of this \link[R6:R6Class]{R6} class.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{RushWorker$new(
  network_id,
  config = NULL,
  worker_id = NULL,
  profile = NULL,
  heartbeat_period = NULL,
  heartbeat_expire = NULL
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{network_id}}{(\code{character(1)})\cr
Identifier of the rush network.
Manager and workers must have the same id.
Keys in Redis are prefixed with the instance id.}
      \item{\code{config}}{(\link[redux:redis_config]{redux::redis_config})\cr
Redis configuration options.
If \code{NULL}, configuration set by \code{\link[=rush_plan]{rush_plan()}} is used.
If \code{rush_plan()} has not been called, the \code{REDIS_URL} environment variable is parsed.
If \code{REDIS_URL} is not set, a default configuration is used.
See \link[redux:redis_config]{redux::redis_config} for details.}
      \item{\code{worker_id}}{(\code{character(1)})\cr
Identifier of the worker.
Keys in redis specific to the worker are prefixed with the worker id.}
      \item{\code{profile}}{(\code{character(1)})\cr
Name of the \code{mirai} compute profile the worker runs on.
If \code{NULL}, the worker runs on the default compute profile.}
      \item{\code{heartbeat_period}}{(\code{integer(1)})\cr
Period of the heartbeat in seconds.
The heartbeat is updated every \code{heartbeat_period} seconds.
Must be at least 1 second.}
      \item{\code{heartbeat_expire}}{(\code{integer(1)})\cr
Time to live of the heartbeat in seconds.
The heartbeat key is set to expire after \code{heartbeat_expire} seconds.
Must be at least \code{heartbeat_period}, otherwise a live worker is reaped as lost between two heartbeats.
Set it larger than the longest pause a worker may experience, for example from garbage collection or swapping,
because a live worker wrongly declared lost can leave a task in an inconsistent state.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RushWorker-pop_task"></a>}}
\if{latex}{\out{\hypertarget{method-RushWorker-pop_task}{}}}
\subsection{\code{RushWorker$pop_task()}}{
  Pop a task from the queue and mark it as running.
Returns \code{NULL} if no task is available.

A worker running on a compute profile takes tasks from the queue of its profile first
and falls back to the shared queue,
so that tasks pushed without a profile are processed by any worker.
A worker running on the default compute profile only takes tasks from the shared queue.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{RushWorker$pop_task(timeout = 1, fields = "xs")}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{timeout}}{(\code{numeric(1)})\cr
Time to wait for task in seconds.}
      \item{\code{fields}}{(\code{character()})\cr
Fields to be returned.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RushWorker-push_running_tasks"></a>}}
\if{latex}{\out{\hypertarget{method-RushWorker-push_running_tasks}{}}}
\subsection{\code{RushWorker$push_running_tasks()}}{
  Create running tasks.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{RushWorker$push_running_tasks(xss, xss_extra = NULL, extra = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{xss}}{(list of named \code{list()})\cr
Lists of arguments for the function e.g. \code{list(list(x1 = 1, x2 = 2), list(x1 = 3, x2 = 4))}.
If \code{xss} is empty, no tasks are created and the method returns an empty \code{character()}.}
      \item{\code{xss_extra}}{(list of named \code{list()})\cr
List of additional information stored along with the task
e.g. \code{list(list(timestamp_xs = Sys.time()), list(timestamp_xs = Sys.time()))}.}
      \item{\code{extra}}{(\code{list})\cr
Deprecated argument for additional information stored along with the task.
Use \code{xss_extra} instead.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    (\code{character()})\cr
Keys of the tasks.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RushWorker-finish_tasks"></a>}}
\if{latex}{\out{\hypertarget{method-RushWorker-finish_tasks}{}}}
\subsection{\code{RushWorker$finish_tasks()}}{
  Save the output of tasks and mark them as finished.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{RushWorker$finish_tasks(keys, yss, yss_extra = NULL, extra = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{keys}}{(\code{character(1)})\cr
Keys of the associated tasks.}
      \item{\code{yss}}{(list of named \code{list()})\cr
Lists of results for the function e.g. \code{list(list(y1 = 1, y2 = 2), list(y1 = 3, y2 = 4))}.}
      \item{\code{yss_extra}}{(list of named \code{list()})\cr
List of additional information stored along with the results
e.g. \code{list(list(timestamp_ys = Sys.time()), list(timestamp_ys = Sys.time()))}.}
      \item{\code{extra}}{(named \code{list()})\cr
Deprecated argument for additional information stored along with the results.
Use \code{yss_extra} instead.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    (\code{RushWorker})\cr
Invisible self.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RushWorker-fail_tasks"></a>}}
\if{latex}{\out{\hypertarget{method-RushWorker-fail_tasks}{}}}
\subsection{\code{RushWorker$fail_tasks()}}{
  Move running tasks to failed and optionally save the condition objects.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{RushWorker$fail_tasks(keys, conditions = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{keys}}{(\code{character()})\cr
Keys of the running tasks to be moved.}
      \item{\code{conditions}}{(\code{list()})\cr
List conditions e.g. \code{list(simpleError("Error"), simpleError("Error"))}.
Defaults to \code{list(message = "Task failed")}.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    (\code{RushWorker})\cr
Invisible self.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RushWorker-set_terminated"></a>}}
\if{latex}{\out{\hypertarget{method-RushWorker-set_terminated}{}}}
\subsection{\code{RushWorker$set_terminated()}}{
  Mark the worker as terminated.
Last step in the worker loop before the worker terminates.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{RushWorker$set_terminated()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    (\code{RushWorker})\cr
Invisible self.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RushWorker-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RushWorker-clone}{}}}
\subsection{\code{RushWorker$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{RushWorker$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
