% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_worker.R
\name{start_worker}
\alias{start_worker}
\title{Start a worker}
\usage{
start_worker(instance_id, ...)
}
\arguments{
\item{instance_id}{(\code{character(1)})\cr
Identifier of the rush instance.}

\item{...}{(\code{any})\cr
Arguments passed to \link[redux:redis_config]{redux::redis_config}.}
}
\description{
Start a worker.
This function initializes the connection to the redis server, loads the packages and globals and calls \link{run_worker}.
Call \verb{$create_worker_script()} of \link{Rush} before calling this function.
}
