% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_worker.R
\name{start_worker}
\alias{start_worker}
\title{Start a worker}
\usage{
start_worker(network_id, ...)
}
\arguments{
\item{network_id}{(\code{character(1)})\cr
Identifier of the rush network.
Controller and workers must have the same instance id.
Keys in Redis are prefixed with the instance id.}

\item{...}{(\code{any})\cr
Arguments passed to \link[redux:redis_config]{redux::redis_config}.}
}
\description{
Starts a worker.
The function is called by the user after creating the worker script with \verb{$create_worker_script()} of \link{Rush}.
The function is started with \verb{Rscript -e 'start_worker(network_id, url, ...)'}.
}
\note{
The function initializes the connection to the Redis data base.
It loads the packages and copies the globals to the global environment of the worker.
The function calls \link{run_worker} to initialize the \link{RushWorker} instance and starts the worker loop.
This function is only called when the worker is started with a script.
}
